Compare the actual etags instead of comparing md5sums
The server side is currently using a shortcut in calculating the values
of the `md5sum` field - it uses the Object Storage (S3) ETag value.

ETag is a MD5. But for the multipart uploaded files, the MD5 is computed
from the concatenation of the MD5s of each uploaded part.

Say you uploaded a 14MB file and your part size is 5MB.
Calculate 3 MD5 checksums corresponding to each part, i.e. the checksum
of the first 5MB, the second 5MB, and the last 4MB.
Then take the checksum of their concatenation.
Since MD5 checksums are hex representations of binary data, just make
sure you take the MD5 of the decoded binary concatenation, not of the
ASCII or UTF-8 encoded concatenation.
When that's done, add a hyphen and the number of parts to get the ETag.